+++ compile-all-examples.ps1
@@ -19,6 +19,7 @@ $examples = @(
     'ErrorHandlingDemo',
     'LongRunningOpDemo',
     'ProgressDemo',
+    'RootCommandDemo',
     'SimpleDemo',
     'SubCommandDemo'
 )